begin to split files apart. Some concern over failure to encapsulate aws client
